backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
websockets>=13.0.0,<17.0.0
langgraph>=0.2.0,<1.0.0
langchain-core>=0.3.0,<1.0.0
langchain-openai>=0.3.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.0.0,<3.0.0
e2b>=1.0.0,<2.0.0
e2b-code-interpreter>=1.0.0,<2.0.0
pydantic-settings>=2.0.0,<3.0.0
PyJWT>=2.8.0,<3.0.0
slowapi>=0.1.9,<1.0.0
structlog>=24.0.0
boto3>=1.35.0
httpx>=0.27.0
